requirements.txt
typer[all]
cryptography
pyyaml
